Reorganize structure of point_of_sale shareds
