+++ app/controllers/concerns/point_of_sale_shared/params_validation.rb
@@ -1,0 +1,8 @@
+module PointOfSaleShared::ParamsValidation
+  extend ActiveSupport::Concern
+
+  CANT_BE_BLANK = "can't be blank"
+
+  include PointOfSaleShared::ParamsValidation::Search
+  include PointOfSaleShared::ParamsValidation::Create
+end
