+++ app/controllers/concerns/point_of_sale_shared/permitted_params.rb
@@ -1,0 +1,6 @@
+module PointOfSaleShared::PermittedParams
+  extend ActiveSupport::Concern
+
+  include PointOfSaleShared::PermittedParams::Search
+  include PointOfSaleShared::PermittedParams::Create
+end
