+++ app/controllers/concerns/point_of_sale_shared/permitted_params/search.rb
@@ -1,0 +1,11 @@
+module PointOfSaleShared::PermittedParams::Search
+  extend ActiveSupport::Concern
+
+  included do
+    private
+
+      def search_point_of_sale_params
+        params.permit(:lat, :lng).to_h
+      end
+  end
+end
